+++ dotnet/src/Nexplore.Practices.Core/IClock.cs
@@ -5,11 +5,12 @@ namespace Nexplore.Practices.Core
     public interface IClock
     {
         DateTime Now { get; }
-
         DateTime UtcNow { get; }
 
-        DateTime Today { get; }
+        DateTimeOffset NowOffset { get; }
+        DateTimeOffset UtcNowOffset { get; }
 
-        DateTime UtcToday { get; }
+        DateOnly Today { get; }
+        DateOnly UtcToday { get; }
     }
 }